+++ src/heuristic.c
@@ -2,9 +2,9 @@
 #define HEURISTIC_GUARD
 #include "heuristic.h"
 
-float heuristic(const Node *node) {
+int heuristic(const Node *node) {
     // TODO: Actual heuristic lol
-    return (float) rand();
+    return (int) rand();
 }
 
 #endif // HEURISTIC_GUARD
